Dispatch all recording tasks
Also provides a way for testing functions to wait on recording tasks
before attempting to get any value. This is necessary so that our
users don't need to manually wait for recording tasks through
Glean.dispatcher.testBlockOnQueue themselves. Another upside of
this is that we don't need to expose Glean.dispatcher.testBlockOnQueue.

Another possibility here would have been to simply use
Glean.dispatcher.testBlockOnQueue inside the testing functions,
but there is always the possibility the queue has a bunch more stuff and
the testing step blocks for too long. I chose the way implemented in
this commit because, although it is a bit more involved it is also cleaner.

A `testing` configuration option is included in this commit, robustness
on that option will come once we work on Bug 1682771.