Add the recommended-requiring-type-checking rules to ts linter
Initially I was simply going to use one specific rule from this set
of recommended rules "require-await", which throws a linter error
when you `await` on an expression which does not return a promise.

That was going to help me remove the `await`s from all the functions
that no longer return promises after the dispatcher changes. It's
importat to make this a rule, because during the dispatcher devolpment
I have run into many hard to find bugs related to awaiting on sync stuff.

Anyways, it turned out this small set of linter rules flagged a bunch of
bugs which we were not aware, thus I left the whole set there.