Provide a way to execute some metrics recording sync
It was inevitable to provide a way to execute recording for specific
metrics without disptaching.

Let's take for example `ping.submit`. That is a dispatched task. When we
submit a ping we always record two metrics: the sequence number and the
start_time. If by any chance we lauched three ping submissions in a row
the order the dispatcher might execute the tasks is:

- ping.submit
- ping.submit
- ping.submit
- seq.add
- start_time.set
- seq.add
- start_time.set
- seq.add
- start_time.set

This would mean that all pings would be sent with the same sequence
number and start_time, since the recordings would only happen after all
pings were submitted.

Intuitively, it seems that all metric recording inside of Glean.js
itself should be done synchronously. I opened Bug 1691037 to verify this
assumption and write documentation about this.