Move away from using path alises
This is in preparation for the next commit.

The issue is: Typescript will not resolve type aliases when compiling code
(see: https://github.com/microsoft/TypeScript/issues/26722#issuecomment-580975983).

Webpack will do that, but you'll that in the next commit
I move away from using webpack for compiling the library.

I personaly hate having the `../../../` everywhere
but unless we are willing to write our own tool
that resolves that on the compiled code, I don't see what else we could do.