Provide the @mozilla/glean package without bundling it
- Why stop bundling?

It makes it easier to provide the type declarations.

- But, how?

If we bundle our code, we need to create type declarations on top of the bundled code,
but when the code is already bundled we lose all the file paths and type richness
(bundled code is Javascript after all), which means, in order to get a rich declaration file
we would need to write it ourselves.

That would be kinda stupid work, since we are using typescript and have all the types information.

- Ok well, create the declaration files based on the unbundled code

Yes sure, that is the obvious solution and is also what this commit prepares for.

Using declaration files from unbundled code on top of bundled code, does not work,
so we need to provide the code in an unbundled format in order for the declaration paths to be correct.

- Observations

The reflex to use webpack and bundle + minify our code, came from my web development days.

I looked at multiple libraries and it is common place to provide the library code unbundled.
Developers will consume that library and then bundle it themselves with their own build system.
When we decide to target static websites, we can provide a bundled version of our code using a CDN,
but that is IMO a bridge we can cross when we get there.

We already also have a bug (https://bugzilla.mozilla.org/show_bug.cgi?id=1693039) to provide
conditional exports for users coming from different environments, but since our first consumers
will be bergamot and rally and both are ok with ES6 modules, we are fine for now.