Add ".js" extensions to all local import statement in lib code
This is a very weird change, I know, but there is good reason behind it:

1. ESM resolution algorithm does not currently support automatic
resolution of file extensions and does not have the hability to
import directories that have an index file. The extension and the name
of the file being import need to _always_ be spcified. See:
https://nodejs.org/api/esm.html#esm_customizing_esm_specifier_resolution_algorithm

2. Typescript does not provide features that change compiled JS code.
This means there is no Typescript feature to include the ".js" on
compiled code. See: https://github.com/microsoft/TypeScript/issues/16577#issuecomment-754941937

All of this results in this weird change. It is the accepted workaround
for this issue and until ES6 imports support automatic resolution of
extensions, we will keep it.