EXTRA: Make webext-ts as ES module and add test command
One very important change in this commit, aside from the obvious
change of adding a command we expected to be there but wasn't "test",
is adding "strip-ansi" as a development dependency.

That is a not a dependecy used by the sample, it is a dependency used by
dependencies of the sample. It is needs to be explicitly declared in
our dependencies to make sure the ESM version of that library is loaded.
Otherwise the CJS dependency is downloaded causes compatibility issues.