Refactor PingUploader to use Dispatcher
This is a long overdue refactoring i.e. it should have been done when
we implemented the dispatcher.

The PingUploader had sort of a mini dispatcher implementation in it,
it would be silly to implement shutdown for this alternate dispatcher
in this PR so I decided to finally refactor it to use the same
dispatcher. The actual Dispatcher class is much more robust as well,
it was implemeted after the Uploader and a lot of thought has gone into
it. I trust it more than the mini dispatcher we used to have inside the
PingUploader.