Dispatcher can't be stopped while it's being shutdown
Stopping the dispatcher while it's was being shutdown
prior to the changes on this commit would cause a never resolving
promise, since the Stop would prevent the dispatcher from reaching
the Shutdown command.

With the changes in this commit, attempting to Stop the dispatcher
while it is being shutdown will actually clear the dispatcher queue.
This way tasks that would be executed after Stopping are not executed
anymore, aside from persistent tasks or the Shutdown task itself.