Stop uploader after executing previous requests
When rate limit was hit, we were enqueueing a Stop command on the
uploader dispatcher. The issue is that the Stop command was a "priority
task" which means it would get enqueued at the front of the queue. This
meant if there were previous requests in the queue, they would be halted too.

The expected behaviour for hitting the rate limit was that requests got
executed up until the point when rate limit was hit and then the
dispatcher could be stopped, halting only subsequent tasks.
This is the behaviour the changes in this commit ensure.