Drop _getWholeStore API from Store impls
I remember we were hesistant to alllow getting the whole stire and that
is the reason why it was put into a special function and the underline
was added. Initially the idea was that it would be used only in tests.
The `get` should then throw an error when it got an empty index.

Turns out not all store implementations (namely webext and tests) were
throwing this error and there were two places in the code (outside of tests)
where we were getting the whole store. These places were the events database,
when getting all store names, and the pings database, when scanning the database.

The cause of the initialization error in Qt was that in the events
database code we were trying to get the whole store by passing an empty
index to the `get` API. Qt being the only spec compliant, was throwing
an error when that happened.