Run the size benchmark against a full main checkout
Instead of switching back and forth between branches this uses an
additional worktree.
This guarantees that the benchmark runs fully against code in the main
branch first, then fully against code in the current checkout.

This requires that the report code in the latest checkout needs to be
able to consume the report from the `main` branch,
so be careful when changing its format.