Use importHelpers compiler flag to make generated Qt lib smaller
importHelpers adds all polyfills necessary for the ES6 target
available on the top of the generated file, instead of defining
the necessary polyfills on top of each module and having a lot
of code repetition. This makes it easier to debug teh generated
file but also makes the final lib considerably smaller.

Since Glean.js does not _need_ all the polyfills this also
adds a bit of dead code to the top of the compiled lib, that is
still better than the redundant mess we had before.