Simplify build config and remove duplication from built lib
Previous to this commit we were building one whole new library for each
platform. That was unnecessary, the only part that is indeed different
per library is the platform/ module implementation.

This commit simplifies that, by building the library only once and
relying ecluisvely on the entry points to make sure each user has access
to the desired platform.

Note 1: The Qt platform will still need to be built separate, that does
not change.

Note 2: This is a breaking change because it breaks the glean_parser
templates, because it changes the metrics and pings import paths. I
could have made this a backwards compatible change, but I think at this
point we are still unstanble enough that I can not do that.