EXTRA: Change log command to accept anything that is loggable
The log command was only accepting strings, which was not ideal because
it made it hard to log errors and stack traces and also console.*
function accept anything so our log command should do the same anyways.

In this commit I also audit all the other places we logged errors
outside of the dispatcher in order to make sure we are now logging
useful stack traces along with error messages.