Serve web sample site from a dev server
Even though this is a static site, we need to serve it from a
development server otherwise the browser will not set the Origin header.
Without that header the telemtry server is not able to allow CORS access
for the request. See: https://bugzilla.mozilla.org/show_bug.cgi?id=1676676#c35