Don't enqueue multiple stop command on uploader dispatcher
The previous behaviour would enqueue a Stop command on the
uploader dispatcher everytime a ping submit happening while
the uploader was throttled or stopped due to other limitations.

That resulted in the dispatcher being immediatelly stopped once
it was resumed due to the throttling window being over, because
the first command it would see would be a stop command.

That could result in long lasting periods of time when the uploader
was just stopped and not processing ping requests.