Get rid of unnecessary rate limiter Stopped state
In the previous implementation we were tying the ability to retry
sending recovarebly failed upload request to the rate limiter state
i.e. a user could only retry after the rate limiter time window was
over. In order to do that we introduced a new state to the rate limiter,
the Stopped state. When max recoverable failures was fit, we would out
the rate limiter in that state. It would work just a like the Throttled
state in the sense that the rate limiter would only get out of it once
the time window was complete.

That is actually not how the implementation in glean-core works. In
there we simply stop the upload worker sending it a Done signal once max
recorable failures is hit. If the worker is re-started immediatelly
after it will be allowed to get new tasks.

That is now how the implementation works in this repository as well. I
wonder if tying the retries to the rate limiter window is still a beter
idea, but that has not caused us problems up until now in glean-core, so
for now, I'll leave it like this.'