Use namespaces instead of singletons where possible
By using namespaces instead of singletons we get rid of a
bit of singleton boilerplate. In an attempt to make the final bundle
smaller I went for this change. In the end the size difference is very
small, but less code is always good so I'll leave this in.'

The only place it still made sense to use singletons was on the Context
module, because that class uses a lot of getters and setters.