Dinamically add metric constructors to metric map
This is nice because it means metrics will work exactly like plugins in
the sense that they are only added to the bundle if they are used.

I am not entirely sure about this change, because it means that if a
given metric is in the database, but was not added to the metric map
(think if a metric is removed from the code, but has not been collected
in a ping once the new code lands) it will be read as invalid by Glean.