Guarantee getMonotonicNow doesn't return a negative value
We were previously extracting the TIME_ORIGIN from Date.now()
in case performance.now was not available in the environment.

That could create negative timestamps, because Date.now is not
guaranteed to be monotonically increasing. (Yeah... it is
contradictiory to have a function called getMonotonicNow, but not
guarantee monotonically increasing timestamps for some platforms.
At least there is a TODO comment to address the incoherence at
some point).

Subtracting the TIME_ORIGIN was essentially pointless and I don't really
know why that was the case before. That function is used by the Timestamp
metric type and for event timestamps which are both about elapsed time.

EXTRA: Address stray TODO comment on TimespanMetricType test suite.