Support the use of an isolated/standalone virtual environment
Prior to this commit, the CLI was strongly opinionated about where the
Python virtual environment should be. Behaviour has now been added to check
if the virtualenv is a subdir of the project dir or not. If it is, we use the
entire absolute virtualenv path as the one to use; if not, we continue with
the original behaviour of assuming the venv is a child dir of the project

There are further improvements we can make, but this at least supports
a standalone virtual environment without breaking existing behaviour