Fix version regex in prepare-release script
The two regexes that are fixed in this commit were missing quotes around the version number. The `sed` commands were failing because of the missing quotes, which in turn meant that the auto-updating docs had old versions. In the QT sample, the version not updating can lead to compilation errors for developers and the error messages are not helpful.