Remove trailing glean.restarted events
When we send a ping, we want to avoid passing unnecessary trailing
restarted events. Before we return the payload, we can step backwards
through our events array and pop each restarted event until we hit an
event that is a non-restarted event. At this point we return the array
since all trailing events have been removed.

In the case that the array does not have a trailing restarted event, we
avoid invoking the new functionality completely.