Port over core functionality for the timing distribution metric
Rewrote the Rust implementation for timing distributions into TS. This
includes a port of the metric itself along with required utilities like
the different types of Histograms.

The generation of histograms is more composition rather than inheritance
so that it can match up more closely to the Rust implementation.

changelog
- port of `timing_distribution` metric
- port of histograms (Linear, Exponential, Functional)
- miscellaneous utilities