Configure the repo for debugging tests
It is not very easy (at least in my case it wasn't) to step through the
unit tests. I added a library `ts-mocha` that allows us to easily set up
configurations for debugging our tests.

I added a `launch.json` file that stores that config so we don't need to
regenerate it every time someone works on this. I also added a few notes
on changing the default timeout length so that tests don't fail when you
are debugging.