Add timing distribution to metrics.ymal
Had to update the constructor of the timing distribution classes so that
the `timeUnit` property came in as a `string` rather than a `TimeUnit`
enum. Using the enum was causing issues when trying to test in the
sample projects as well as the schema unit tests.

Other metrics that use a time unit do the same thing and use a string
rather than an explicit type. Everywhere we use the unit it still gets
cast to `TimeUnit` so it doesn't change anything functionally.

changelog
- make constructor param for time unit a string instead of `TimeUnit`
- add entry to metrics.ymal for timing distribution
- start and stop a timer in the unit test for the glean schema