Add count to DistributionData
There isn't an easy way to total up the number of entries in the
histogram with what we provide. Since it doesn't cost us anything
computationally, we can return the count in the snapshot so that a
user can easily see how many entries are in the histogram and
verify that all of their timings got logged.