Update dependabot PR frequency
Now that we are getting more up-to-date, we likely aren't going to need
weekly checks. We can move this to monthly and on the first of the month
we can review and merge the PRs. It is currently on weekly, which is
causing it to just open too many PRs.