Skip .spec files in tests/unit/platform/utils/webext/sample
Without explicitly skipping `.spec` files in this directory we get an
error with how node is loading modules. This is related to our use of
experimental ESM modules.

This is the cleanest, least-hacky way to ensure that we are able to
ignore this unnecessary file. Since the spec file is inside of the node
module, it has already been compiled to JS code and is not even used.
Another alternative is to just delete the file using something like
`patch-package`, but that is a bit too hacky for this scenario.